Limpeza e organização do código
