+++ demo-rest/src/main/java/com/keeggo/demo_rest/Main.java
@@ -9,5 +9,4 @@ public class Main {
 	public static void main(String[] args) throws IOException, InterruptedException {
 		Application.main(args);
 	}
-	//teste Github
 }
