--- demo-rest/src/main/resources/features/login.feature
@@ -1,11 +1,0 @@
-#language: pt
-Funcionalidade: Login
-
-  @CT02_Fazer_login
-  Esquema do Cenário: 12021532369260 - Fazer login na api
-    Dado que faço login na api com o nome "<NOME>" e senha "<SENHA>"
-
-    Exemplos: 
-      | NOME              | SENHA    |
-      | admin@default.com | Rsi@2020 |
-
